[decompose] extract Zap and LinkPreview controllers from VideoModal.js
