+++ docs/agents/task-logs/daily/2026-02-20_decompose-agent_completed.md
@@ -1,0 +1,27 @@
+# Task Completion Log: Decompose Agent
+
+**Date:** 2026-02-20
+**Agent:** decompose-agent
+**Status:** Completed
+
+## Summary
+Successfully decomposed `js/ui/components/VideoModal.js` by extracting distinct functional blocks into separate controllers. This improved code organization and reduced the file size significantly.
+
+### Decomposition Details
+- **Source File:** `js/ui/components/VideoModal.js` (Lines before: 6045)
+- **Extracted Controllers:**
+  1.  `js/ui/components/video-modal/zapController.js`: Encapsulates all Zap dialog logic (popover, form, receipts, visibility).
+  2.  `js/ui/components/video-modal/linkPreviewController.js`: Encapsulates link preview fetching and rendering logic.
+- **Lines Reduced:** 1069 (New size: 4976 lines)
+
+### Bug Fixes
+- Addressed code review feedback: Updated `ZapController` and `LinkPreviewController` to access `document` and `window` dynamically from the modal instance to avoid initialization timing issues.
+
+### Verification
+- **Lint:** Passed (`npm run lint`).
+- **Unit Tests:** Passed (`npm run test:unit`).
+- **Baseline:** Updated `scripts/check-file-size.mjs` with new line count for `js/ui/components/VideoModal.js`.
+
+### Artifacts
+- `context/CONTEXT_2026-02-20_decompose-agent.md`: Detailed plan and rationale.
+- New files: `js/ui/components/video-modal/zapController.js`, `js/ui/components/video-modal/linkPreviewController.js`.
